+++ doc/changelog.md
@@ -8,6 +8,7 @@ This files contains a version history including all changes relevant for semanti
 ## Upcoming Changes
 
 * Fix: multiple Controller registrations for same Interface
+* **new Feature**: Added json-based Endpoint-Info (http-GET on controller-route) and corresponding evaluation-features at client side to retrieve version-information